최종본

+++ TodoList/src/main/webapp/resources/js/detail.js
@@ -31,8 +31,20 @@ completeBtn.addEventListener("click",()=>{
 // 삭제 버튼
 const deleteBtn = document.querySelector("#deleteBtn");
 deleteBtn.addEventListener("click",()=>{
+
+
+  if(!confirm("정말 삭제하시겠습니까?"))return;
+
   location.href="/todo/delete?todoNo="+todoNo;
 
 });
 // 수정 버튼
+
+const updateBtn = document.querySelector("#updateBtn");
+updateBtn.addEventListener("click",()=>{
+  //수정할 수 있는 화면을 요청(GET 방식)
+  location.href="/todo/update?todoNo="+todoNo;
+
+
+});
